# api-swagger.yaml
openapi: 3.0.3
info:
  title: OpenMemory API
  description: |
    OpenMemory is a memory management system with advanced features including:
    - Memory storage and retrieval
    - Temporal fact tracking
    - Reflection and insights
    - IDE integration
    - Compression utilities
    - Dynamic salience calculations
  version: 2.0.0
  contact:
    name: OpenMemory
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.openmemory.ai
    description: Production server
...
paths:
...
  /dashboard/stats:
    get:
      tags:
        - Dashboard
      summary: Get dashboard statistics
      responses:
        '200':
          description: Dashboard stats
        '500':
          description: Internal server error
...
  /dashboard/activity:
    get:
      tags:
        - Dashboard
      summary: Get recent activity
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Activity list
        '500':
          description: Internal server error
  /dashboard/sectors/timeline:
    get:
      tags:
        - Dashboard
      summary: Get sector timeline
      parameters:
        - name: hours
          in: query
          schema:
            type: integer
            default: 24
      responses:
        '200':
          description: Sector timeline
        '500':
          description: Internal server error
  /dashboard/top-memories:
    get:
      tags:
        - Dashboard
      summary: Get top memories by salience
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Top memories
        '500':
          description: Internal server error
  /dashboard/maintenance:
    get:
      tags:
        - Dashboard
      summary: Get maintenance operations
      parameters:
        - name: hours
          in: query
          schema:
            type: integer
            default: 24
      responses:
        '200':
          description: Maintenance operations
        '500':
          description: Internal server error
...